Member 칼럼 money, Goods 칼럼 price 를 int -> long 으로 형변환

+++ src/main/java/com/leehw/sbly/config/auth/Dto/SessionMember.java
@@ -10,7 +10,7 @@ public class SessionMember implements Serializable {
     private Long id;
     private String name;
     private String email;
-    private int money;
+    private long money;
 
     public SessionMember(Member member){
         this.id = member.getId();
